UI test cleanup: Extract explicit_counter_loop tests
